Restrict access with controller action

+++ app/controllers/articles_controller.rb
@@ -25,23 +25,37 @@ class ArticlesController < ApplicationController
   end
 
   def edit
+    unless @article.user == current_user
+      flash[:alert] = "You can only edit your own article"
+      redirect_to root_path
+    end
   end
   
   def update
-    if @article.update(article_params)
-      flash[:success] = "Article has been updated"
-      redirect_to @article
-    else
-      flash.now[:danger] = "Article has not been updated"
-      render :edit
-    end
+    unless @article.user == current_user
+      flash[:danger] = "You can only edit your own article"
+      redirect_to root_path
+    else 
+      if @article.update(article_params)
+        flash[:success] = "Article has been updated"
+        redirect_to @article
+      else
+        flash.now[:danger] = "Article has not been updated"
+        render :edit
+      end
+    end  
   end
 
   def destroy
-    if @article.destroy
-      flash[:success] = "Article has been deleted"
-      redirect_to articles_path
-    end
+    unless @article.user == current_user
+      flash[:alert] = "You can only delete your own article"
+      redirect_to root_path
+    else
+      if @article.destroy
+        flash[:success] = "Article has been deleted"
+        redirect_to articles_path
+      end
+    end  
   end
